Move Event/EventObserver classes from feature modules to :core:ui
