+++ core/ui/src/main/java/com/chaeny/busoda/ui/event/Event.kt
@@ -1,4 +1,4 @@
-package com.chaeny.busoda.stoplist.event
+package com.chaeny.busoda.ui.event
 
 /**
  * Used as a wrapper for data that is exposed via a LiveData that represents an event.
