+++ core/ui/src/main/java/com/chaeny/busoda/ui/event/EventObserver.kt
@@ -1,4 +1,4 @@
-package com.chaeny.busoda.stopdetail.event
+package com.chaeny.busoda.ui.event
 
 import androidx.lifecycle.Observer
