+++ feature/stopdetail/build.gradle
@@ -54,5 +54,6 @@ dependencies {
 
     implementation project(":core:data")
     implementation project(":core:model")
+    implementation project(":core:ui")
     testImplementation project(":core:testing")
 }
