+++ feature/stopdetail/src/main/java/com/chaeny/busoda/stopdetail/StopDetailFragment.kt
@@ -8,7 +8,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.chaeny.busoda.stopdetail.databinding.FragmentStopDetailBinding
-import com.chaeny.busoda.stopdetail.event.EventObserver
+import com.chaeny.busoda.ui.event.EventObserver
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
