--- feature/stopdetail/src/main/java/com/chaeny/busoda/stopdetail/event/Event.kt
@@ -1,27 +1,0 @@
-package com.chaeny.busoda.stopdetail.event
-
-/**
- * Used as a wrapper for data that is exposed via a LiveData that represents an event.
- */
-open class Event<out T>(private val content: T) {
-
-    var hasBeenHandled = false
-        private set
-
-    /**
-     * Returns the content and prevents its use again.
-     */
-    fun getContentIfNotHandled(): T? {
-        return if (hasBeenHandled) {
-            null
-        } else {
-            hasBeenHandled = true
-            content
-        }
-    }
-
-    /**
-     * Returns the content, even if it's already been handled.
-     */
-    fun peekContent(): T = content
-}
