--- feature/stoplist/src/main/java/com/chaeny/busoda/stoplist/event/EventObserver.kt
@@ -1,11 +1,0 @@
-package com.chaeny.busoda.stoplist.event
-
-import androidx.lifecycle.Observer
-
-class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
-    override fun onChanged(event: Event<T>) {
-        event.getContentIfNotHandled()?.let { value ->
-            onEventUnhandledContent(value)
-        }
-    }
-}
